+++ dev/js-WebAssembly/Network.js
@@ -739,7 +739,7 @@ class Network {
     }
 
     static get version () {
-        return "3.3.3"
+        return "3.3.4"
     }
 }