+++ test/js-test.js
@@ -36,7 +36,7 @@ describe("Loading", () => {
     })
 
     it("Statically returns the Network version when accessing via .version", () => {
-        expect(Network.version).to.equal("3.3.1")
+        expect(Network.version).to.equal("3.3.4")
     })
 })